Añadido nuevas alertas, cambios en diseño en Login-Register, Eliminación de archivos

+++ client/src/components/main/changeboards/SideBarBoards.js
@@ -1,15 +1,17 @@
 import React from 'react';
 import { useDispatch, useSelector } from 'react-redux';
+import Swal from 'sweetalert2';
 
 import useClickOutside from '../../../utils/functions/useClickOutside';
+import { DivCloseIconAbsolute } from '../../../utils/components/Div';
+import { ButtonComment } from '../../../utils/components/Button';
+import { ToastTimer } from '../../../utils/alerts/Alert';
+import { H3 } from '../../../utils/components/Titles';
 import api from '../../../redux/action-creator';
 import actions from '../../../redux/actions';
-import { DivCloseIconAbsolute } from '../../../utils/components/Div';
-import { H3 } from '../../../utils/components/Titles';
 
-import sContainer from '../../../styles/container.module.css';
 import HighlightOffIcon from '@material-ui/icons/HighlightOff';
-import { ButtonComment } from '../../../utils/components/Button';
+import sContainer from '../../../styles/container.module.css';
 
 export default function SideBarBoards({ sideBar, setSideBar }) {
 
@@ -34,37 +36,60 @@ export default function SideBarBoards({ sideBar, setSideBar }) {
   }
 
   const onHandleDelete = (el) => {
-    const data = { uuid: el.uuid, idUser: user.id }
+
     if (user.dashboards.length > 1) {
-        let position = user.dashboards.findIndex(el2 => el2.uuid === el.uuid)
-        dispatch(api.deleteDashboard(data))
-
-        user.dashboards.splice(position, 1)
-        dispatch({ type: USER, payload: user })
-
-        dispatch({ type: DASHBOARD, payload: user.dashboards[position === 0 ? position ++ : 0] })
-        dispatch(api.getColumn(user.dashboards[position === 0 ? position : 0].uuid))
-
-        setSideBar(false)
-        return 
+      Swal.fire({
+        title: 'Are you sure?',
+        text: "You won't be able to revert this!",
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonColor: '#d33',
+        confirmButtonText: 'Yes, delete it!'
+      }).then((result) => {
+        if (result.isConfirmed) {
+          const data = { uuid: el.uuid, idUser: user.id }
+          let position = user.dashboards.findIndex(board => board.uuid === el.uuid)
+          dispatch(api.deleteDashboard(data))
+          user.dashboards.splice(position, 1)
+          dispatch({ type: USER, payload: user })
+          dispatch({ type: DASHBOARD, payload: user.dashboards[position === 0 ? position++ : 0] })
+          dispatch(api.getColumn(user.dashboards[position === 0 ? position : 0].uuid))
+          ToastTimer.fire(
+            'Deleted!',
+            'Your board has been deleted.',
+            'success'
+          )
+        }
+      })
     } else {
-        return alert("No puedes eliminar el ultimo proyecto")
+      ToastTimer.fire(
+        'Deny!',
+        "you can't delete de last board",
+        "warning"
+      )
     }
-}
+  }
 
   return (
     <div ref={domnNode} className={sContainer.sideBarBoards}>
       <DivCloseIconAbsolute onClick={onHandleClose} />
       <H3 title={"Boards"} s={"titleBoards"} style={{ textAlign: "center" }} />
-      <hr/>
+      <hr />
       <div className={sContainer.containerBoards}>
         {user.dashboards?.map(el => {
           let some = el.uuid === dashboard.uuid
           return <div
             key={el.uuid}
             className={some ? sContainer.containerMapBoardsUuid : sContainer.containerMapBoards}>
-            <ButtonComment label={el.title} onClick={onHandleSelectProject} el={el} s={"buttonBoards"}/>
-            <HighlightOffIcon className={sContainer.iconDeleteBoard} onClick={() => onHandleDelete(el)}/>
+            <ButtonComment
+              label={el.title}
+              onClick={onHandleSelectProject}
+              el={el}
+              s={"buttonBoards"} />
+            <HighlightOffIcon
+              className={sContainer.iconDeleteBoard}
+              onClick={() => onHandleDelete(el)}
+            />
           </div>
         })}
       </div>
